reactor: use thread cputime clock for stall detection
Using the thread cputime clock prevents false positives due to an overcommitted
processor, and this improves the quality of the reports.

It could also generate false negatives if system calls block (since the cputime
clock is stopped when a thread switches contexts voluntarily), but these events
are rarer these days and are better investigated using the other latency detector.