Add a dev build mode.
We currently have two modes, but they are both slow to build:

* release is slow to build because of optimizations.
* debug is slow to build because of debug info and sanitizers.

This patch adds a dev mode. It is useful as a first target when
iterating on a patch when all that is needed is a fast build and
running a few tests.

Before posting a patch one is still expected to run tests in release
and/or debug mode.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>