build: Use absolute shared library paths for Boost
When Boost is compiled locally as part of a development environment, we
want to ensure that all shared libraries reference their dependencies on
other Boost shared libraries with absolute paths in the build tree. This
is because run-time dependencies of the libraries when linked to
executables need to reflect the local Boost and not the system-installed
one.