tests: Define and use `SEASTAR_TESTING_MAIN`
We are restricted to dynamic linking with the Boost Test library since
we need to provide a custom entry point.

The Boost library conditionally declares `main` based on the definition
of `BOOST_TEST_DYN_LINK`, which we previously defined unconditionally as
part of the implementation of the test runner.

Now, authors of Seastar test programs can conditionally define
`SEASTAR_TESTING_MAIN` to include the definition of a default entry
point which simply invokes `seastar::entry_point`.

If `SEASTAR_TESTING_MAIN` is not defined, then a test program will need
to define it manually (and likely invoke `seastar::entry_point`).

We still require that `BOOST_TEST_DYN_LINK` is unconditionally defined,
but we do it in the build system rather than the source file.