reactor: harden io_getevents() callsite against signals
Currently, io_getevents() never enters the kernel due to the way we call it.
However, we plan to add a debug option to in fact enter the kernel; when that
is active, the system call can fail with EINTR, triggering an assertion failure.

Detect this condition and prevent the assertion failure. We treat EINTR as a timeout;
it should always be safer to return early from io_getevents().