testing/seastar_test: throw exception when no tests are registered
Currently if there are no tests registered (because e.g. the developer
forgot to compile the .cc file containing the tests into the test
executable) we dereference an invalid pointer and crash with SEGFAULT in
the best case. This is confusing and will possibly cause the poor soul
who caused this to try to find the non-existing problem in their
test(s). To avoid this add a check and throw an exception when there are
no tests are registered, with a message that will hopefully make it
clear what is the problem.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <e76d3638155208bc991384006ef3c0254d7efed8.1545139915.git.bdenes@scylladb.com>