build: Re-run Doxygen unconditionally
Previously, a marker file was used in combination with a dependency on
the compiled Seastar library to only run Doxygen when source files had
changed.

Instead, we unconditionally re-run Doxygen when the `doc_api` target is
built. That target is not built by default, and executing Doxygen is a
relatively fast process.

Fixes #558

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <96ec725371b811f6bab3c6362b5d961cb4093861.1545238416.git.jhaberku@scylladb.com>